requirements.txt
ace_tools==0.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloc==0.2.5
colored==2.3.0
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.57.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
pyan3==1.2.0
pydeps==3.0.1
pyparsing==3.2.3
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==59.6.0
shiboken6==6.9.0
six==1.17.0
stdlib-list==0.11.1
tomli==2.2.1
tzdata==2025.2
urllib3==2.4.0

